# Faster batchnorm inference