--- a/nestedtensor/csrc/autograd_functions.cpp
+++ b/nestedtensor/csrc/autograd_functions.cpp
@@ -134,7 +134,7 @@ Tensor NestedTensor_batch_norm(
     Tensor output = input;
     output = NestedTensor_contiguous(output);
     Tensor input_buffer = get_buffer(output);
-    Tensor output_buffer = input_buffer.clone();
+    // Tensor output_buffer = input_buffer.clone();
   
     auto self_opt_sizes = get_opt_sizes(input);
   
@@ -170,13 +170,15 @@ Tensor NestedTensor_batch_norm(
         c10::Half((float)(eps)),
         weight_ptr,
         bias_ptr,
-        output_buffer.data_ptr<c10::Half>(),
+        input_buffer.data_ptr<c10::Half>(),
+        // output_buffer.data_ptr<c10::Half>(),
         (int)(*self_opt_sizes[0] * *self_opt_sizes[1]),
         (int)(*self_opt_sizes[0]),
         nt_sizes.data_ptr<int>(),
         defaultStream
         );
-    return wrap_buffer(std::move(output_buffer), get_efficient_nested_size(output), get_efficient_nested_stride(output));
+    // return wrap_buffer(std::move(output_buffer), get_efficient_nested_size(output), get_efficient_nested_stride(output));
+    return wrap_buffer(std::move(input_buffer), get_efficient_nested_size(output), get_efficient_nested_stride(output));
   }
 #endif