--- a/nestedtensor/csrc/cuda/add.cu
+++ b/nestedtensor/csrc/cuda/add.cu
@@ -144,7 +144,7 @@ void sub_scalar_kernelLauncher(
 
 __global__
 void batchnorm_inference(
-    c10::Half* input,
+    const c10::Half* input,
     c10::Half* mean,
     c10::Half* running_var,
     c10::Half eps,
@@ -155,26 +155,26 @@ void batchnorm_inference(
     const int* offsets)
 {
   const int batch_id  = blockIdx.x;
+  const int grid_id  = blockIdx.y;
   const int scalars_id  = batch_id / input_outer_stride;
-  const int grain_size = blockDim.x;
-  const int tid = threadIdx.x;
+  const int grain_size = 256 * 2;
+  const int tid = threadIdx.x + grid_id * 256;
   const int range = (offsets[batch_id + 1] - offsets[batch_id]);
   const int num_chunks = range / grain_size;
   c10::Half value = running_var[scalars_id] + eps;
   value = hrsqrt(value);
   value = value * weight[scalars_id];
-  for (int id = 0; id < num_chunks; id++) {
-    output[offsets[batch_id] + id * grain_size + tid] =
-      (((input[offsets[batch_id] + id * grain_size + tid] - mean[scalars_id])
-       * value)
-       + bias[scalars_id]);
+  c10::Half value2 = mean[scalars_id] * value - bias[scalars_id];
+
+  int input_offset = offsets[batch_id] + tid;
+  int id = 0;
+  // printf("num_chunks: %d\n", num_chunks);
+  for (; id < num_chunks; id++) {
+    output[input_offset] = __ldg(reinterpret_cast<const __half*>(input) + input_offset) * value - value2;
+    input_offset += grain_size;
   }
-  const int leftover = num_chunks * grain_size;
-  if (leftover + tid < range) {
-    output[offsets[batch_id] + leftover + tid] =
-      (((input[offsets[batch_id] + leftover + tid] - mean[scalars_id])
-       * value)
-       + bias[scalars_id]);
+  if (input_offset < offsets[batch_id + 1]) { //leftover + tid < range) {
+    output[input_offset] = __ldg(reinterpret_cast<const __half*>(input) + input_offset) * value - value2;
   }
 }
 
@@ -193,6 +193,7 @@ void batchnorm_inference_kernelLauncher(
 {
   dim3 grid;
   grid.x = batch_size;
+  grid.y = 2;
 
   batchnorm_inference<<<grid, 256, 0, stream>>>(
       input,