--- a/nestedtensor/version.py
+++ b/nestedtensor/version.py
@@ -1,5 +1,5 @@
-__version__ = '0.1.4+40b4a63'
-git_version = '40b4a637ed257b0cc6dc09bd87b0508735ef4015'
+__version__ = '0.1.4+e56fbdb'
+git_version = 'e56fbdb0ae2974fdf8b3d4f1c5648ffc86de487e'
 from nestedtensor import _C
 if hasattr(_C, 'CUDA_VERSION'):
     cuda = _C.CUDA_VERSION